[wip]signup リファクタ　続き
