+++ php.log
@@ -1,32 +1,87 @@
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：アカウント作成ページ
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 画面表示処理開始
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッションID：49ri22597q391ib7qniu40g8gi
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッション変数の中身：Array
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：アカウント作成ページ
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 画面表示処理開始
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：セッションID：ho1hocb3qo4gt8f12039svo6d5
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：セッション変数の中身：Array
 (
     [msg_success] => 
-    [login_date] => 1609851753
+    [login_date] => 1610026222
     [login_limit] => 3600
     [user_id] => 0
+    [test] => パスワード(再入力)が合っていません
 )
 
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：現在日時タイムスタンプ：1609922362
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：ログイン期限日時タイムスタンプ：1609855353
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッション変数の中身：
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：アカウント作成ページ
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 画面表示処理開始
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッションID：dekmahr48o1htpjh1dkt245qlj
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッション変数の中身：Array
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：現在日時タイムスタンプ：1610026940
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：ログイン期限日時タイムスタンプ：1610029822
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：ユーザー登録処理に入りました。
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：classes\admin\signup Object
+(
+    [email:protected] => 
+    [pass:protected] => password
+    [password_re:protected] => 
+    [err_ms:protected] => 
+)
+
+[07-Jan-2021 22:42:20 Asia/Tokyo] PHP Notice:  Undefined index: user_login in /Applications/MAMP/htdocs/EmRev/EmRev_app/index.php on line 96
+[07-Jan-2021 22:42:20 Asia/Tokyo] デバッグ：セッション変数の中身：
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：アカウント作成ページ
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 画面表示処理開始
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：セッションID：834oo6fq1bfu1u40701fi1i27n
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：セッション変数の中身：Array
 (
     [msg_success] => 
-    [login_date] => 1609851753
+    [login_date] => 1610026222
     [login_limit] => 3600
     [user_id] => 0
+    [test] => パスワード(再入力)が合っていません
 )
 
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：現在日時タイムスタンプ：1609922362
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：ログイン期限日時タイムスタンプ：1609855353
-[06-Jan-2021 17:39:22 Asia/Tokyo] デバッグ：セッション変数の中身：
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：現在日時タイムスタンプ：1610026990
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：ログイン期限日時タイムスタンプ：1610029822
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：ユーザー登録処理に入りました。
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：classes\admin\signup Object
+(
+    [email:protected] => 
+    [pass:protected] => password
+    [password_re:protected] => 
+    [err_ms] => 
+)
+
+[07-Jan-2021 22:43:10 Asia/Tokyo] PHP Notice:  Undefined index: user_login in /Applications/MAMP/htdocs/EmRev/EmRev_app/index.php on line 96
+[07-Jan-2021 22:43:10 Asia/Tokyo] デバッグ：セッション変数の中身：
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：アカウント作成ページ
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 画面表示処理開始
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：セッションID：13ba335t5m3pud440vttda88nr
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：セッション変数の中身：Array
+(
+    [msg_success] => 
+    [login_date] => 1610026222
+    [login_limit] => 3600
+    [user_id] => 0
+    [test] => パスワード(再入力)が合っていません
+)
+
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：現在日時タイムスタンプ：1610026991
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：ログイン期限日時タイムスタンプ：1610029822
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「「「「「「「
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：ユーザー登録処理に入りました。
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：「「「「「「「「「「「「「
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：classes\admin\signup Object
+(
+    [email:protected] => 
+    [pass:protected] => password
+    [password_re:protected] => 
+    [err_ms] => 
+)
+
+[07-Jan-2021 22:43:11 Asia/Tokyo] PHP Notice:  Undefined index: user_login in /Applications/MAMP/htdocs/EmRev/EmRev_app/index.php on line 96
+[07-Jan-2021 22:43:11 Asia/Tokyo] デバッグ：セッション変数の中身：
